feat: User defined default runtime
Allows users to define the Ant framework default runtime by adding the "defaultRuntime" entry in the configuration file.
Its value should be the name of the runtime to be chosen as default.

Also, the user can now specify which type of function will be created when using "function add" command, by using the "--type" option.