fix: --config option related bugs (#63)
* fix: Plugins basePath not being resolved

The relative basePaths of plugins were not resolved, resulting in errors during Ant's execution. With this fix, now the basePaths are resolved accordingly with Ant's basePath.

* fix: CLI saving configuration relative paths

All of our CLI commands that adds entries into the configuration file are not resolving the relative paths input by the user.
In the configuration file, the paths are relative to the basePath entry, not to the current working directory. In the CLI, usually when the user inputs a relative path, it is relative to the current working directory.
To avoid mistakes when adding new configuration entries, the solution was to always resolve the CLI commands relative paths and make it save only absolute paths into the configuration file.

* fix: GraphQL model file not being found when using --config

Taking the Ant configuration basePath into consideration resolves this. That way, the GraphQL plugin can resolve its basePath and avoid using the process.cwd.