refactor: Analytics as class
Refactor the `Analytics` module to a class, making it easier to write tests.
Changes were done to the `initialize` function as well, now receiving only one parameter, a configuration object which contains all properties needed to initialize the APIs used for tracking.
Some of the `Analytics` function are now asynchronous, instead of returning Promises.
This commit also introduces the `AntStorageController` which overrides Parse's default storage controller. This is required in order to keep track of the installation ID from Parse, allowing us to track the users actions with greater precision.