fix: GraphQL API is not working for operations that return an array
As solution, now the `resolve.js` (which represents the `@resolve` directive) flattens the `Observable` returned from the function invoked.
This solution fails in the case where the function returns an `Observable` that emmits arrays, but in this case, the `Observable` should be handled in order to return the expected type of data.