feat: Support for runtimes with multiple versions
Added support for runtimes with multiple versions.
Now runtimes must have the `version` attribute defined in order to be supported by the framework.
The `runtime add`, `runtime remove` CLI commands now requires the `version` argument, in order to manipulate the correct runtime instance.
The `function add` command can now receive the `runtimeVersion` argument, which allows users to specify the version of the runtime that will execute the created function.
The `runtime ls` CLI command shows the runtime version and which runtimes are the defaults.
The `RuntimeController` API has changed as well. The `getRuntime` function can now receive the `version` parameter, which when defined, tries to retrieve the runtime with this specific version. If not defined, returns the default version of the given runtime `name`.