feat: Error tracking (#76)
* feat: Introducing Analytics

Introduces analytics into the framework by tracking the command ran by the users. The request itself is done by a separate process, in order to avoid blocking the main process.

* fix: Avoid tracking events when running tests

Mocks were needed when running tests from Core and GraphQL plugin modules. Refactoring of both modules are needed in order to remove the `@back4app/ant-cli` dependency and thus the need of having the `@back4app/ant-util-analytics` mock.

* fix: dependencies versions

* test: increase coverage

* feat: Introducing error tracking

Introduces Sentry as our error tracking tool.
Refactor was needed to adjust all yargs error handlers, in order to avoid exiting the process before finishing the tracking process.
The timeout of 2000ms is set to send the tracking request to Sentry servers.

* refactor: Analytics initialization

Refactors the `ant-util-analytics` module in order to make its initialization optional. Now the initialization is done by the `Analytics.initialize` function, which receives both Parse and Sentry information needed for their initializations.
This way, no mocks are needed to avoid tracking unecessary events, like tests runs, for instance.

* refactor: Analytics as class

Refactor the `Analytics` module to a class, making it easier to write tests.
Changes were done to the `initialize` function as well, now receiving only one parameter, a configuration object which contains all properties needed to initialize the APIs used for tracking.
Some of the `Analytics` function are now asynchronous, instead of returning Promises.
This commit also introduces the `AntStorageController` which overrides Parse's default storage controller. This is required in order to keep track of the installation ID from Parse, allowing us to track the users actions with greater precision.

* fix: Avoid tracking handled/expected errors

The purpose of error tracking is to identify any bugs or incorrect behaviors as soon as possible and provide a solution for it. Handled errors are expected (such as incorrect typing of commands), and we should not consider them as major problems.